Major refactor: Overhaul entire codebase architecture
- Restructured project layout for improved modularity
- Implemented clean separation of concerns across modules
- Refactored database operations into a cohesive API
- Overhauled test suite to use public APIs
- Updated CMakeLists.txt to reflect new project structure
