+++ src/client.h
@@ -1,6 +1,7 @@
 #ifndef CLIENT_H
 #define CLIENT_H
 
+#include "handler.h"
 #include "resp.h"        // Include this for the Parser definition
 #include "ring_buffer.h" // Include this if ring_buffer is defined in a separate header
 
@@ -12,5 +13,9 @@ typedef struct Client {
   ring_buffer output_buffer;
   struct Parser *parser;
 } Client;
+
+Client *create_client(int fd, Handler *handler);
+void flush_client_output(Client *client);
+void destroy_client(Client *client);
 
 #endif // CLIENT_H
