+++ src/command_handler.h
@@ -9,7 +9,7 @@ extern "C" {
 #include <stddef.h> // For size_t
 #include <stdint.h> // For int64_t
 
-struct Handler;
+typedef enum { CMD_PING, CMD_ECHO, CMD_SET, CMD_GET, CMD_UNKNOWN } CommandType;
 
 typedef struct CommandHandler {
   char *buf;
