+++ src/server.h
@@ -13,32 +13,7 @@ extern "C" {
 struct CommandHandler;
 
 // type definitions
-typedef enum { TYPE_STRING } ValueType;
 
-typedef struct {
-  ValueType type;
-  union {
-    char *str;
-  } data;
-  time_t expiration;
-} RedisValue;
-
-typedef struct {
-  long long expiration; // expiration time in milliseconds sinch epoch
-  int nx;               // flag for NX option
-  int xx;               // flag for XX option
-  int keepttl;          // flag for KEEPTTL option
-  int get;              // flag for get option
-} SetOptions;
-
-KHASH_MAP_INIT_STR(redis_hash, RedisValue *)
-
-long long current_time_millis();
-void set_value(khash_t(redis_hash) * h, const char *key, const void *value, ValueType type,
-               long long expiration);
-RedisValue *get_value(khash_t(redis_hash) * h, const char *key);
-void cleanup_hash(khash_t(redis_hash) * h);
-void handle_command(struct CommandHandler *ch);
 int start_server();
 
 #ifdef __cplusplus
